third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-ice-promise.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the RTCPeerConnection Ice functionality.");

var pc = null;
var iceCandidate = null;

function onIceChange2()
{
    if (pc.iceConnectionState === "closed") {
        testPassed("iceConnectionState is closed.");
        finishJSTest();
    }
}

function addIceCandidateSuccess()
{
    testPassed("addIceCandidateSuccess was called.");
    pc.oniceconnectionstatechange = onIceChange2;
    pc.close();
}

function addIceCandidateFailure()
{
    testFailed("addIceCandidateFailure was called.");
    finishJSTest();
}

function unexpectedSuccess()
{
    testFailed("unexpectedSuccess was called.");
    finishJSTest();
}

function expectedTypeError(error)
{
  window.error = error;
  shouldBe('error.name', '"TypeError"')
  testPassed('expectedTypeError was called.')
}

function onIceChange1()
{
    if (pc.iceConnectionState === "completed") {
        testPassed("iceConnectionState is completed");
        iceCandidate = new RTCIceCandidate({candidate:"nano nano"});
        shouldNotThrow('pc.addIceCandidate(null).then(unexpectedSuccess, expectedTypeError);');
        shouldNotThrow('pc.addIceCandidate(iceCandidate).then(addIceCandidateSuccess, addIceCandidateFailure);');
    }
}

shouldNotThrow('pc = new RTCPeerConnection();');
pc.oniceconnectionstatechange = onIceChange1;

window.jsTestIsAsync = true;
window.successfullyParsed = true;
</script>
</body>
</html>
