third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-ice-promise-expected.txt

Tests the RTCPeerConnection Ice functionality.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc = new RTCPeerConnection(); did not throw exception.
PASS iceConnectionState is completed
PASS pc.addIceCandidate(null).then(unexpectedSuccess, expectedTypeError); did not throw exception.
PASS pc.addIceCandidate(iceCandidate).then(addIceCandidateSuccess, addIceCandidateFailure); did not throw exception.
PASS error.name is "TypeError"
PASS expectedTypeError was called.
PASS addIceCandidateSuccess was called.
PASS iceConnectionState is closed.
PASS successfullyParsed is true

TEST COMPLETE
